style: 폴더명 api 명에 맞게 변경. 카카오 로그인 위해
