--- src/app/auth/login/page.tsx
@@ -1,13 +1,0 @@
-import React from "react";
-import Login from "./Login";
-
-function Main() {
-  return (
-    <div>
-      <h1>My App</h1>
-      <Login />
-    </div>
-  );
-}
-
-export default Main;
