--- src/app/auth/onboarding/page.tsx
@@ -1,13 +1,0 @@
-import React from "react";
-import { RecoilRoot } from "recoil";
-import type { AppProps } from "next/app";
-
-function Home({ Component, pageProps }: AppProps) {
-  return (
-    <RecoilRoot>
-      <Component {...pageProps} />
-    </RecoilRoot>
-  );
-}
-
-export default Home;
