--- src/app/auth/onboarding/sub/page.css
@@ -1,9 +1,0 @@
-.button-field {
-  width: 375px;
-  height: 76px;
-}
-.button {
-  width: 76px;
-  height: 48px;
-  background-color: 999999;
-}
