--- src/app/auth/onboarding/sub/page.tsx
@@ -1,12 +1,0 @@
-import React from "react";
-import Email from "./email";
-
-function Main() {
-  return (
-    <div>
-      <Email />
-    </div>
-  );
-}
-
-export default Main;
