+++ src/app/page.tsx
@@ -1,12 +1,13 @@
 import React from "react";
-// import SignIn from "./auth/login/page";
-import Email from "./auth/onboarding/sub/email";
+// import Start from "./auth/onboarding/start/page";
+import { RecoilRoot } from "recoil";
+import Login from "./member/login/Start";
 
 function Home() {
   return (
-    <main>
-      <Email />
-    </main>
+    <RecoilRoot>
+      <Login />
+    </RecoilRoot>
   );
 }
